+++ tests/test_core_io_buffered.py
@@ -1,6 +1,7 @@
 # coding=utf-8
 """Test pycosio._core.io_buffered"""
 import io
+import os
 import time
 
 import pytest
@@ -69,6 +70,7 @@ def test_object_buffered_base_io():
         _RAW_CLASS = DummyRawIO
         DEFAULT_BUFFER_SIZE = buffer_size
         MINIMUM_BUFFER_SIZE = 10
+        MAXIMUM_BUFFER_SIZE = 10000
 
         def ensure_ready(self):
             """Ensure flush is complete"""
@@ -90,6 +92,26 @@ def test_object_buffered_base_io():
             """Flush"""
             self._write_futures.append(self._workers.submit(
                 flush, self._write_buffer[:self._buffer_seek]))
+
+    class DummyRawIOPartFlush(DummyRawIO):
+        """Dummy IO with part flush support"""
+        _SUPPORT_PART_FLUSH = True
+        _size = 20
+
+        def _flush(self, buffer, start, *_):
+            """Do nothing"""
+            if start == 50:
+                # Simulate buffer that need to wait previous one
+                time.sleep(0.1)
+            raw_flushed.extend(buffer)
+
+    class DummyBufferedIOPartFlush(ObjectBufferedIOBase):
+        """Dummy buffered IO with part flush support"""
+        _RAW_CLASS = DummyRawIOPartFlush
+
+    class DummyBufferedIONoPartFlush(ObjectBufferedIOBase):
+        """Dummy buffered IO without part flush support"""
+        _RAW_CLASS = DummyRawIO
 
     # Test raw
     object_io = DummyBufferedIO(name)
@@ -245,3 +267,18 @@ def test_object_buffered_base_io():
     assert object_io._buffer_size == 1000
     object_io = DummyBufferedIO(name, mode='w', buffer_size=1)
     assert object_io._buffer_size == DummyBufferedIO.MINIMUM_BUFFER_SIZE
+    object_io = DummyBufferedIO(name, mode='w', buffer_size=1000000)
+    assert object_io._buffer_size == DummyBufferedIO.MAXIMUM_BUFFER_SIZE
+
+    # Test raises not implemented error if no part flush support and no
+    # _flush implementation
+    object_io = DummyBufferedIONoPartFlush(name, mode='w')
+    with pytest.raises(NotImplementedError):
+        object_io.flush()
+
+    # Test default implementation with part flush support
+    raw_flushed[:] = b''
+    content = os.urandom(100)
+    with DummyBufferedIOPartFlush(name, mode='w', buffer_size=10) as object_io:
+        object_io.write(content)
+    assert raw_flushed == content